Code-quality sweep: zero clippy warnings
- clippy --fix applied for style lints (needless iterator iteration,
  explicit clone closures, str::replace chains, split_once, map().flatten(),
  match-for-equality, etc.)
- Removed dead orphaned spec_decode::router() + its now-unused axum imports.
- Added type alias LmStudioRow to factor out the 7-tuple query type.
- Marked intentional future-feature dead code with #[allow(dead_code)] +
  notes: MetacognitiveScore, OwlFamilyCoverage (owl metacognitive pass),
  canonicalize (used in unit tests).
- Allowed clippy::too_many_arguments on spawn_run_task (cohesive arg set).

Verified: cargo clippy --release -> 0 warnings; cargo build --release clean;
backend serves; modular run regression (run 758) still works.

+++ src/routes/fountain.rs
@@ -238,7 +238,7 @@ async fn run_probe(
         }
     }
 
-    let sent = (ok_n + limited_n + err_n) as i64;
+    let sent = ok_n + limited_n + err_n;
     let verdict = verdict_for(ok_n, sent, limited_n);
     let duration = started.elapsed().as_millis() as i64;
     // Seal the full per-request evidence — same provenance discipline as runs.
